Copy File Change

+++ Framework/Built_In_Automation/Built_In_Utility/CrossPlatform/BuiltInUtilityFunction.py
@@ -878,14 +878,29 @@ def raw(text):
 def Copy_File_or_Folder(step_data):
     sModuleInfo = inspect.stack()[0][3] + " : " + inspect.getmoduleinfo(__file__).name
     CommonUtil.ExecLog(sModuleInfo, "Function start", 0)
+
+    # Recall file attachment, if not already set
+    file_attachment = []
+    if Shared_Resources.Test_Shared_Variables('file_attachment'):
+        file_attachment = Shared_Resources.Get_Shared_Variables('file_attachment')
+
     try:
         if _platform == "linux" or _platform == "linux2" or _platform == "darwin":
-            from_path = get_home_folder() + str(step_data[0][2]).strip()  # location of the file/folder to be copied
+            from_path = str(step_data[0][2]).strip()  # location of the file/folder to be copied
             to_path = get_home_folder() + str(step_data[1][2]).strip()  # location where to copy the file/folder
         elif _platform == "win32":
             from_path = raw(str(step_data[0][2]).strip())  # location of the file/folder to be copied
             to_path = raw(str(step_data[1][2]).strip())  # location where to copy the file/folder
         file_or_folder = str(step_data[2][2]).strip()  # get if it is file/folder to copy
+        # Try to find the file
+        if from_path not in file_attachment and os.path.exists(os.path.join(get_home_folder(), from_path)) == False:
+            CommonUtil.ExecLog(sModuleInfo,
+                               "Could not find file attachment called %s, and could not find it locally" % from_path, 3)
+            return 'failed'
+        if from_path in file_attachment: from_path = file_attachment[from_path]  # In file is an attachment, get the full path
+
+        if from_path not in file_attachment:
+            file_name = os.path.join(get_home_folder(), from_path)
         if file_or_folder.lower() == 'file':
                 # copy file "from_path" to "to_path"
             result = copy_file(from_path, to_path)
